Parallelised Berger-Parker's diversity index

This function computes the Berger-Parker diversity index in a parallelised manner, improving performance on suitable hardware for large datasets.

Usage

BergerParkerP(rasterm, w, na.tolerance, debugging)

Arguments

rasterm

Input raster data as a matrix or suitable object convertible to a matrix, containing the values on which to compute the index.

w

Integer, representing half of the side length of the square moving window used in calculations.

na.tolerance

Numeric value between 0.0 and 1.0 indicating the proportion of NA values tolerated in each moving window over the input rasterm. If the proportion of NA's exceeds na.tolerance, the result for that window is set as NA. The default is 0.0, indicating no tolerance for NA values.

debugging

Logical, when TRUE, additional console messages are printed for debugging purposes. Defaults to FALSE.

Value

A matrix or a list of matrices with the Berger-Parker index computed through a moving window of the specified size.

Details

Parallelised Berger-Parker's diversity index

See also

Author

Marcantonio Matteo marcantoniomatteo@gmail.com

Martina Iannacito martina.iannacito@inria.fr

Duccio Rocchini duccio.rocchini@unibo.it

Examples

if (FALSE) {
# Generate a sample matrix
data_matrix <- matrix(sample(c(NA, 1:10), 100, replace = TRUE), 10, 10)

# Compute the Berger-Parker index with parallel processing
result <- BergerParkerP(rasterm = data_matrix, w = 1, na.tolerance = 0.1, debugging = TRUE)
}