Renyi's entropy (H)

Computes Renyi's entropy ($$H$$) on different classes of numeric matrices using a moving window algorithm.

Usage

Renyi(
  x,
  window = 3,
  alpha = 1,
  base = exp(1),
  rasterOut = TRUE,
  np = 1,
  na.tolerance = 1,
  cluster.type = "SOCK",
  debugging = FALSE
)

Arguments

x

Input data, which may be a matrix, a Spatial Grid Data Frame, a SpatRaster, or a list of these objects. In the latter case, only the first element of the list will be considered.

window

The side of the square moving window; it must be an odd numeric value greater than 1 to ensure that the target pixel is in the centre of the moving window. Default value is 3.

alpha

Order of diversity to compute the index. If alpha is a vector with length greater than 1, then the index will be calculated over x for each value in the sequence.

base

A numerical value defining the base of the logarithm in Renyi's entropy formula. Default is exp(1).

rasterOut

Boolean; if TRUE, the output will be in SpatRaster format with x as a template.

np

The number of processes (cores) to be spawned. Default is 1.

na.tolerance

A numeric value between 0.0 and 1.0 indicating the proportion of NA values tolerated when calculating Renyi's index in each moving window over x. If the relative proportion of NAs is larger than na.tolerance, the value for that window is set as NA; otherwise, Renyi's index is calculated considering the non-NA values. Default is 1.0 (no tolerance for NAs).

cluster.type

The type of cluster to create. Options are "MPI" (calls "makeMPIcluster"), "FORK", and "SOCK" (calls "makeCluster"). Default is "SOCK".

debugging

Boolean; if TRUE, additional messages will be printed for debugging purposes. Default is FALSE.

Value

A list of matrices with a length equal to the length of alpha. If the length of alpha is 1, then a matrix of dimensions dim(x) is returned.

Details

Renyi's entropy ($$H$$) is calculated on a numerical matrix as $$H = \frac{1}{(1-q)} \ln(\sum_{i=1}^{R} {p^q}_i)$$, where q is the considered order of diversity (alpha), R is the total number of categories (i.e., unique numerical values in the considered numerical matrix), and p is the relative abundance of each category. If q=1, Shannon.R is called to calculate $$H'$$ instead of the indefinite $${}^1D$$. If $$p > 2*10^9$$, then BerkgerParker.R is called to calculate $$log(1/{}^\infty H)$$. Renyi's entropy of low order gives more weight to rarer categories, while higher orders weight more dominant categories.

Note

For Linux users: to enable MPI parallel computing, install libopenmpi by running the following commands in the terminal: apt-get update; apt-get upgrade; apt-get install mpi; apt-get install libopenmpi-dev; apt-get install r-cran-rmpi.

Windows users might require additional steps to use "MPI". Refer to the detailed guide at: https://bioinfomagician.wordpress.com/2013/11/18/installing-rmpi-mpi-for-r-on-mac-and-windows/

References

Rényi, A. (1970). Probability Theory. North Holland Publishing Company, Amsterdam.

See also

Author

Matteo Marcantonio marcantoniomatteo@gmail.com, Martina Iannacito martina.iannacito@inria.fr, Duccio Rocchini duccio.rocchini@unibo.it

Examples

# Minimal example: compute Renyi's index with alpha 1:5 
a <- matrix(c(10,10,10,20,20,20,20,30,30), ncol=3, nrow=3)
renyi <- Renyi(x = a, window = 3, alpha = 1:5)
#> 
#> Object x check OK: 
#> Renyi with alpha parameter value in  1 2 3 4 5  will be returned.
#> 
#> Processing alpha 1
#> Progress:  50 on 100  Progress:  75 on 100  Progress: 100 on 100  
#> 
#> Processing alpha 2
#> 
#> 
#> Processing alpha 2 Window 3
#> 
#> Processing alpha 3
#> 
#> 
#> Processing alpha 3 Window 3
#> 
#> Processing alpha 4
#> 
#> 
#> Processing alpha 4 Window 3
#> 
#> Processing alpha 5
#> 
#> 
#> Processing alpha 5 Window 3
#> 
#> Calculation complete.