Parallelised Renyi's diversity index

Parallelised Renyi's diversity index.

Usage

RenyiP(rasterm, w, alpha, base, na.tolerance, debugging)

Arguments

rasterm

input data.

w

half of the side of the square moving window.

alpha

alpha value for order of diversity in Renyi's Index.

base

base of the logarithm to calculate the index.

na.tolerance

a numeric value \((0.0-1.0)\) which indicates the proportion of NA values that will be tolerated to calculate Rao's index in each moving window over x. If the relative proportion of NA's in a moving window is bigger than na.tolerance, then the value of the window will be set as NA, otherwise Rao's index will be calculated considering the non-NA values. Default values is 0.0 (i.e., no tolerance for NA's).

debugging

a boolean variable set to FALSE by default. If TRUE, additional messages will be printed. For de-bugging only.

Value

Matrix or a list of matrixes with the Renyi index computed through moving window of the given size.

Author

Marcantonio Matteo marcantoniomatteo@gmail.com
Martina Iannacito martina.iannacito@inria.fr
Duccio Rocchini duccio.rocchini@unibo.it

See also