Sequential Parametric Rao's index of quadratic entropy (Q)

Computes the sequential version of the parametric Rao's index of quadratic entropy (Q), a measure used in environmental and ecological studies to assess biodiversity by considering the evolutionary distance between species. The function performs calculations in a sequential manner over a moving window across the input data.

Usage

paRaoS(
  x,
  alpha,
  window,
  dist_m,
  na.tolerance,
  diag,
  debugging,
  isfloat,
  mfactor
)

Arguments

x

Matrix or data frame; the input data over which the index calculation is performed.

alpha

Numeric; specifies the alpha value for the order of diversity in Hill's Index.

window

Numeric; half of the side length of the square moving window used in the calculation.

dist_m

Character; specifies the type of distance metric used in calculations.

na.tolerance

Numeric; the threshold proportion of NA values allowed in the moving window. If exceeded, the calculation for that window is skipped. Values range from 0.0 (no tolerance) to 1.0.

diag

Logical; indicates whether the diagonal of the distance matrix should be included in the computation. Typically set to FALSE.

debugging

Logical; set to FALSE by default. If TRUE, additional console messages will be displayed for debugging purposes.

isfloat

Logical; indicates whether the input data values are floating-point numbers.

mfactor

Integer; indicates the decimal position to round.

Value

A list of matrices corresponding to the computed Rao's index values. Each matrix in the list represents the calculations performed over the moving window, with dimensions equal to dim(x).

See also

Author

Duccio Rocchini duccio.rocchini@unibo.it, Matteo Marcantonio marcantoniomatteo@gmail.com