Pielou's evenness index (E')

Computes Pielou's evenness index on different classes of numeric matrices using a moving window algorithm.

Usage

Pielou(
  x,
  window = 3,
  rasterOut = TRUE,
  np = 1,
  na.tolerance = 1,
  cluster.type = "SOCK",
  debugging = FALSE
)

Arguments

x

Input data may be a matrix, a Spatial Grid Data Frame, a SpatRaster or a list of these objects. In the latter case, only the first element of the list will be considered.

window

The side of the square moving window, it must be an odd numeric value greater than 1 to ensure that the target pixel is in the centre of the moving window. Default value is 3.

rasterOut

Boolean, if TRUE the output will be in SpatRaster format with x as template.

np

The number of processes to use for parallel processing. Defaults to 1.

na.tolerance

A numeric value (0.0-1.0) indicating the proportion of NA values tolerated for calculating Pielou's index in each moving window over x. Defaults to 1.0 (i.e., full tolerance).

cluster.type

Specifies the type of cluster to create. Available options are "MPI" (utilizing "makeMPIcluster"), "FORK", and "SOCK" (using "makeCluster"). Default is "SOCK".

debugging

A boolean; if TRUE, additional debugging messages will be printed. Defaults to FALSE.

Value

A numerical matrix with the same dimensions as x.

Details

Pielou evenness's index (E') is calculated as E' = $${\sum_{i=1}^{R} p_i \times log(p_i)}\over{log(R)}$$, where R represents the total number of categories (i.e., unique numerical values) and p is the relative abundance of each category. The index represents the ratio between the observed diversity and the maximum possible diversity.

Note

Linux users need to install libopenmpi for MPI parallel computing. For Linux Ubuntu, try the following commands: apt-get update; apt-get upgrade; apt-get install mpi; apt-get install libopenmpi-dev; apt-get install r-cran-rmpi Windows users may need to perform additional setup for "MPI". More information is available at: https://bioinfomagician.wordpress.com/2013/11/18/installing-rmpi-mpi-for-r-on-mac-and-windows/

References

Pielou, E.C. (1966). The measurement of diversity in different types of biological collections. Journal of Theoretical Biology, 13: 131-144.

See also

Examples

# Minimal example; compute Pielou's index  
a <- matrix(c(10,10,10,20,20,20,20,30,30), ncol=3, nrow=3)
renyi <- Pielou(x=a, window=3)
#> 
#> Object x check OK: 
#> Pielou output matrix will be returned.
#> Progress:  50 on 100  Progress:  75 on 100  Progress: 100 on 100  
#> 
#> Calculation complete.